+++ src/components/TextInput/TextInput.vue
@@ -1,12 +1,13 @@
 <template>
   <div :class="{ 'relative': withCopyButton }">
-    <lob-label
+    <LobLabel
       v-if="label"
       :label="label"
       :label-for="id"
       :required="required"
       :sr-only-label="srOnlyLabel"
       :tooltip-content="tooltipContent"
+      :tooltip-position="tooltipPosition"
     />
     <div
       v-if="withCopyButton"
@@ -114,7 +115,8 @@
       </div>
       <button
         v-if="withCopyButton"
-        class="rounded-full px-3 h-7 type-xs-700 bg-black text-white hover:bg-gray-700 focus:outline-dashed focus:outline-black focus:outline-offset-1 active:bg-gray-800 focus:bg-gray-800"
+        type="button"
+        class="rounded-full px-3 h-7 type-xs-700 bg-black text-white hover:bg-gray-700 focus-visible:outline-dashed focus-visible:outline-black focus-visible:outline-offset-1 active:bg-gray-800 focus:bg-gray-800"
         @click="copyToClipboard"
       >
         {{ copyButtonLabel }}
@@ -147,6 +149,10 @@ export default {
       type: String,
       default: null
     },
+    tooltipPosition: { type: String, default: 'trailing',
+      validator: function (value) {
+        return ['leading', 'trailing'].includes(value);
+      } },
     modelValue: {
       type: [String, Number],
       default: null